refactor: extract 3 local-filter-state panels from PanelViews.tsx, fixes #245
Batch 4 of the per-panel split for issue #144: schools, users, content
- the panels each owning their own filter/search local state instead of
sharing values from usePanelData.

- New panels/SchoolsPanel.tsx ({ schools } prop) - moves in the
  stateFilter/distFilter/blockFilter useState + filteredSchools
  derivation, previously top-level in PanelViews.tsx and used only here.
- New panels/UsersPanel.tsx ({ usersList } prop) - moves in
  userRoleFilter/userSearch useState, used only here.
- New panels/ContentPanel.tsx (no props - fully self-contained via
  FLN_LEVELS_LIST) - moves in its own search/classFilter useState.
  This is issue #245's fix: those two hooks used to live inside
  PanelViews.tsx's conditional if (panel === 'content') block (a React
  hooks-order violation that crashed the whole app when navigating here
  from any other panel, since PanelViews never unmounts). As this
  component's own top-level hooks they're now called unconditionally
  every render, closing the bug as a side effect of the mechanical move.
- PanelViews.tsx's 3 if-blocks replaced with one-line router calls.
- Removed the now-fully-orphaned top-level search/setSearch (line 41)
  along with the filter state above - confirmed via grep this was
  already dead even before this PR (shadowed and never read at the top
  level, only content's own inner copy was ever used).
- Unused-import cleanup (same audit method as prior batches): removed
  BookMarked, FLN_LEVELS_LIST import, STATE_NAMES, BLOCK_NAMES - all
  confirmed dead, their only usages were in the 3 panels moved here.
  Kept DISTRICT_NAMES (still used by districts/blocks panels).

Verified: tsc --noEmit clean, vite build succeeds, byte-diff of every
moved block. Live click-through as Superadmin: reproduced #245's exact
crash steps (navigate to Analytics, then Content) and confirmed it no
longer happens - Content renders all 93 levels correctly, console
clean of any React/hooks errors. Schools panel's state filter verified
to actually filter (1440 -> 40 schools for Andhra Pradesh). Users
panel verified with its default role filter (5 of 6449 shown).
